+++ backend/blog_api/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 from blog_api.models import Blog
-from blog_api.models import BlogComment
+# from blog_api.models import BlogComment
 
 
 
@@ -14,10 +14,10 @@ class BlogAdmin(admin.ModelAdmin):
     readonly_fields = ('id',)
 
 
-@admin.register(BlogComment)
-class BlogCommentAdmin(admin.ModelAdmin):
-    list_display = ('id', 'blog', 'commentor', 'comment', 'is_deleted')
-    list_filter = ('commentor',)
-    search_fields = ('comment', )
-    readonly_fields = ('id', )
+# @admin.register(BlogComment)
+# class BlogCommentAdmin(admin.ModelAdmin):
+#     list_display = ('id', 'blog', 'commentor', 'comment', 'is_deleted')
+#     list_filter = ('commentor',)
+#     search_fields = ('comment', )
+#     readonly_fields = ('id', )
