+++ backend/blog_api/urls/v1.py
@@ -9,5 +9,5 @@ urlpatterns = [
     path('<int:pk>', views.BlogView.as_view()),
     path('archive/<int:pk>', views.BlogArchiveView.as_view()),
     path('unarchive/<int:pk>', views.BlogUnarchiveView.as_view()),
-    path('comment', views.BlogCommentView.as_view()),
+    # path('comment', views.BlogCommentView.as_view()),
 ]
